Add setAppOpsUidMode
On Android P or even earlier, uid mode rather than package mode is used for runtime permissions. To modify uid mode, setUidMode should be used.

This commit adds setAppOpsUidMode in the client which does not set EXTRA_PACKAGE_NAME, so the manager can use this to judge is setUidMode should be used.